+++ yandex/data_source_yandex_function_test.go
@@ -99,6 +99,10 @@ func TestAccDataSourceYandexFunction_full(t *testing.T) {
 	}
 	params.zipFilename = "test-fixtures/serverless/main.zip"
 	params.maxAsyncRetries = "2"
+	params.logOptions = testLogOptions{
+		disabled: false,
+		minLevel: "WARN",
+	}
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:     func() { testAccPreCheck(t) },
@@ -130,6 +134,9 @@ func TestAccDataSourceYandexFunction_full(t *testing.T) {
 					resource.TestCheckResourceAttr(functionDataSource, "storage_mounts.0.prefix", params.storageMount.storageMountPrefix),
 					resource.TestCheckResourceAttr(functionDataSource, "storage_mounts.0.read_only", fmt.Sprint(params.storageMount.storageMountReadOnly)),
 					resource.TestCheckResourceAttr(functionDataSource, "async_invocation.0.retries_count", params.maxAsyncRetries),
+					resource.TestCheckResourceAttr(functionResource, "log_options.0.disabled", fmt.Sprint(params.logOptions.disabled)),
+					resource.TestCheckResourceAttr(functionResource, "log_options.0.min_level", params.logOptions.minLevel),
+					resource.TestCheckResourceAttrSet(functionResource, "log_options.0.log_group_id"),
 					testAccCheckCreatedAtAttr(functionDataSource),
 				),
 			},
@@ -221,6 +228,11 @@ resource "yandex_function" "test-function" {
   async_invocation {
     retries_count = "%s"
   }
+  log_options {
+  	disabled = "%t"
+	log_group_id = yandex_logging_group.logging-group.id
+	min_level = "%s"
+  }
 }
 
 resource "yandex_resourcemanager_folder_iam_member" "sa-editor" {
@@ -230,6 +242,9 @@ resource "yandex_resourcemanager_folder_iam_member" "sa-editor" {
 }
 
 resource "yandex_iam_service_account_static_access_key" "sa-static-key" {
+  depends_on = [
+	yandex_resourcemanager_folder_iam_member.sa-editor,
+  ]
   service_account_id = yandex_iam_service_account.test-account.id
 }
 
@@ -262,6 +277,9 @@ resource "yandex_lockbox_secret_version" "secret_version" {
     text_value = "%s"
   }
 }
+
+resource "yandex_logging_group" "logging-group" {
+}
 	`,
 		params.name,
 		params.desc,
@@ -281,6 +299,8 @@ resource "yandex_lockbox_secret_version" "secret_version" {
 		params.storageMount.storageMountReadOnly,
 		params.zipFilename,
 		params.maxAsyncRetries,
+		params.logOptions.disabled,
+		params.logOptions.minLevel,
 		params.storageMount.storageMountBucket,
 		params.serviceAccount,
 		params.secret.secretName,